+++ cmd/curator/main.go
@@ -22,7 +22,7 @@ import (
 )
 
 const (
-	version = "0.21.0"
+	version = "0.22.0"
 )
 
 func main() {